requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
exceptiongroup==1.1.1
fastapi==0.95.1
firebase-admin==6.1.0
Flask==2.2.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
msgpack==1.0.5
mysql-connector-python==8.0.32
numpy==1.24.2
opencv-python==4.7.0.72
outcome==1.2.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pyfcm==1.5.4
PyJWT==2.6.0
pynput==1.7.6
pyparsing==3.0.9
pypng==0.20220715.0
PySocks==1.7.1
qrcode==7.4.2
requests==2.28.2
rsa==4.9
selenium==3.141.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.26.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wsproto==1.2.0